refactor: 코드 로직 수정
- 프로토콜과 구현체를 별도의 파일로 분리해두던 것을 합침
- 불필요한 if let 구문을 해제함
