--- Tekken8 Frame Data/TK8/Repository/Data/MoveRepository.swift
@@ -1,10 +1,0 @@
-//
-//  MoveRepository.swift
-//  Tekken8 Frame Data
-//
-//  Created by 문영균 on 4/16/25.
-//
-
-protocol MoveRepository {
-    func fetchMoves(characterName name: String) async throws -> [Move]
-}
